Make the Construction Run

+++ include/TRsignal.h
@@ -1,6 +1,7 @@
 #include <vector>
 #include <utility>
 #include <stdio.h>
+#include <cstdint>
 
 #ifndef TRSIGNAL_H
 #define TRSIGNAL_H
@@ -10,10 +11,10 @@ class Signal{
 
 public:
     
-    std::vector<int> signal;
+    std::vector<uint8_t> signal;
 
-    Signal(const std::vector<int>& signal) : signal(signal) { };
-    Signal(std::vector<int>&& signal) : signal(std::move(signal)) { };
+    Signal(const std::vector<uint8_t>& signal) : signal(signal) { };
+    Signal(std::vector<uint8_t>&& signal) : signal(std::move(signal)) { };
 
     int len() { return signal.size(); };
